Work on Smart link
