+++ tailwind-client/src/component/Modals/NavbarModals/AnnouncementModal/LinkPreview.module.css
@@ -1,0 +1,46 @@
+.view_container {
+  -ms-overflow-style: none; /* IE and Edge */
+  scrollbar-width: none;
+}
+
+/* Hide scrollbar for Chrome, Safari and Opera */
+.view_container::-webkit-scrollbar {
+  display: none;
+}
+
+.textarea_container {
+  overflow: auto;
+  outline: none;
+  border: 0;
+
+  -webkit-box-shadow: none;
+  -moz-box-shadow: none;
+  box-shadow: none;
+
+  resize: none;
+}
+
+.textarea_container::-webkit-scrollbar {
+  display: none;
+}
+
+.textarea_container:focus {
+  box-shadow: none;
+}
+
+.container {
+  background: rgb(238, 174, 202);
+  background: radial-gradient(circle, rgba(238, 174, 202, 1) 0%, rgba(148, 187, 233, 1) 100%);
+}
+
+.card_container {
+  background: rgb(57, 57, 57);
+  background: linear-gradient(146deg, rgba(57, 57, 57, 1) 0%, rgba(0, 0, 0, 1) 65%);
+}
+
+.card_header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background-color: antiquewhite;
+}
