--- tailwind-client/src/component/Modals/NavbarModals/PopModals.module.css
@@ -1,24 +1,0 @@
-.view_container {
-  -ms-overflow-style: none; /* IE and Edge */
-  scrollbar-width: none;
-}
-
-/* Hide scrollbar for Chrome, Safari and Opera */
-.view_container::-webkit-scrollbar {
-  display: none;
-}
-
-.textarea_container {
-  overflow: auto;
-  outline: none;
-
-  -webkit-box-shadow: none;
-  -moz-box-shadow: none;
-  box-shadow: none;
-
-  resize: none;
-}
-
-.textarea_container::-webkit-scrollbar {
-  display: none;
-}
