+++ tailwind-client/src/component/Modals/NavbarModals/index.js
@@ -1,0 +1,3 @@
+export { default as AnnouncementModal } from './AnnouncementModal/AnnouncementModal';
+export { default as UploadVideoModal } from './UploadVideoModal/UploadVideoModal';
+export { default as UploadTrackModal } from './UploadTrackModal/UploadTrackModal';
